requirements.txt
sentence-transformers
qdrant-client
openai
fastapi
uvicorn
streamlit
wikipedia
python-dotenv
